Virtual Agent Grid Lis

+++ TrinityService/src/main/java/com/netpro/trinity/service/agent/controller/VRAgentController.java
@@ -18,19 +18,29 @@ import org.springframework.web.bind.annotation.RestController;
 import com.netpro.trinity.service.agent.entity.VRAgent;
 import com.netpro.trinity.service.agent.service.VRAgentService;
 import com.netpro.trinity.service.dto.FilterInfo;
+import com.netpro.trinity.service.permission.feign.PermissionClient;
+import com.netpro.trinity.service.util.ACUtil;
 
 @RestController  //宣告一個Restful Web Service的Resource
 @RequestMapping("/vragent")
 public class VRAgentController {
 	private static final Logger LOGGER = LoggerFactory.getLogger(VRAgentController.class);
-		
+	
 	@Autowired
 	private VRAgentService service;
 	
+	@Autowired
+	private PermissionClient permissionClient;
+	
 	@GetMapping("/findAll")
-	public ResponseEntity<?> findAllVRAgents(Boolean withoutDetail) {
+	public ResponseEntity<?> findAllVRAgents(HttpServletRequest request, Boolean withoutDetail) {
 		try {
-			return ResponseEntity.ok(this.service.getAll(withoutDetail));
+			String peopleId = ACUtil.getUserIdFromAC(request);
+			if(this.permissionClient.checkFuncPermission(peopleId, "agent", "view")) {
+				return ResponseEntity.ok(this.service.getAll(withoutDetail));
+			}else {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You do not have 'View' Permission!");
+			}
 		}catch(Exception e) {
 			VRAgentController.LOGGER.error("Exception; reason was:", e);
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
@@ -38,22 +48,14 @@ public class VRAgentController {
 	}
 	
 	@GetMapping("/findByUid")
-	public ResponseEntity<?> findVRAgentByUid(Boolean withoutDetail, String uid) {
+	public ResponseEntity<?> findVRAgentByUid(HttpServletRequest request, Boolean withoutDetail, String uid) {
 		try {
-			return ResponseEntity.ok(this.service.getByUid(withoutDetail, uid));
-		}catch(IllegalArgumentException e) {
-			VRAgentController.LOGGER.error("IllegalArgumentException; reason was:", e);
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
-		}catch(Exception e) {
-			VRAgentController.LOGGER.error("Exception; reason was:", e);
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
-		}
-	}
-  
-	@GetMapping("/findByName")
-	public ResponseEntity<?> findVRAgentsByName(Boolean withoutDetail, String name) {
-		try {
-			return ResponseEntity.ok(this.service.getByName(withoutDetail, name));
+			String peopleId = ACUtil.getUserIdFromAC(request);
+			if(this.permissionClient.checkPermission(peopleId, "agent", uid, "view")) {
+				return ResponseEntity.ok(this.service.getByUid(withoutDetail, uid));
+			}else {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You do not have 'View' Permission!");
+			}
 		}catch(IllegalArgumentException e) {
 			VRAgentController.LOGGER.error("IllegalArgumentException; reason was:", e);
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
@@ -64,9 +66,14 @@ public class VRAgentController {
 	}
 	
 	@PostMapping("/findByFilter")
-	public ResponseEntity<?> findVRAgentsByFilter(Boolean withoutDetail, @RequestBody FilterInfo filter) {
+	public ResponseEntity<?> findVRAgentsByFilter(HttpServletRequest request, Boolean withoutDetail, @RequestBody FilterInfo filter) {
 		try {
-			return this.service.getByFilter(withoutDetail, filter);
+			String peopleId = ACUtil.getUserIdFromAC(request);
+			if(this.permissionClient.checkFuncPermission(peopleId, "agent", "view")) {
+				return this.service.getByFilter(withoutDetail, filter);
+			}else {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You do not have 'View' Permission!");
+			}
 		}catch(SecurityException e) {
 			VRAgentController.LOGGER.error("SecurityException; reason was:", e);
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
@@ -91,7 +98,12 @@ public class VRAgentController {
 	@PostMapping("/add")
 	public ResponseEntity<?> addVRAgent(HttpServletRequest request, @RequestBody VRAgent vragent) {
 		try {
-			return ResponseEntity.ok(this.service.add(request, vragent));
+			String peopleId = ACUtil.getUserIdFromAC(request);
+			if(this.permissionClient.checkFuncPermission(peopleId, "agent", "add")) {
+				return ResponseEntity.ok(this.service.add(request, vragent));
+			}else {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You do not have 'Add' Permission!");
+			}
 		}catch(IllegalArgumentException e) {
 			VRAgentController.LOGGER.error("IllegalArgumentException; reason was:", e);
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
@@ -102,9 +114,14 @@ public class VRAgentController {
 	}
 	
 	@PostMapping("/edit")
-	public ResponseEntity<?> editVRAgent(@RequestBody VRAgent vragent) {
+	public ResponseEntity<?> editVRAgent(HttpServletRequest request, @RequestBody VRAgent vragent) {
 		try {
-			return ResponseEntity.ok(this.service.edit(vragent));
+			String peopleId = ACUtil.getUserIdFromAC(request);
+			if(this.permissionClient.checkPermission(peopleId, "agent", vragent.getVirtualagentuid(),  "modify")) {
+				return ResponseEntity.ok(this.service.edit(vragent));
+			}else {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You do not have 'Edit' Permission!");
+			}
 		}catch(IllegalArgumentException e) {
 			VRAgentController.LOGGER.error("IllegalArgumentException; reason was:", e);
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
@@ -115,9 +132,14 @@ public class VRAgentController {
 	}
   
 	@GetMapping("/delete")
-	public ResponseEntity<?> deleteVRAgentByUid(String uid) {
+	public ResponseEntity<?> deleteVRAgentByUid(HttpServletRequest request, String uid) {
 		try {
-			this.service.deleteByUid(uid);
+			String peopleId = ACUtil.getUserIdFromAC(request);
+			if(this.permissionClient.checkPermission(peopleId, "agent", uid,  "delete")) {
+				this.service.deleteByUid(uid);
+			}else {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You do not have 'Delete' Permission!");
+			}
 		}catch(IllegalArgumentException e) {
 			VRAgentController.LOGGER.error("IllegalArgumentException; reason was:", e);
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
@@ -129,9 +151,14 @@ public class VRAgentController {
 	}
 	
 	@GetMapping("/isExistByUid")
-	public ResponseEntity<?> isVRAgentExistByUid(String uid) {
+	public ResponseEntity<?> isVRAgentExistByUid(HttpServletRequest request, String uid) {
 		try {
-			return ResponseEntity.ok(this.service.existByUid(uid));
+			String peopleId = ACUtil.getUserIdFromAC(request);
+			if(this.permissionClient.checkFuncPermission(peopleId, "agent", "view")) {
+				return ResponseEntity.ok(this.service.existByUid(uid));
+			}else {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You do not have 'View' Permission!");
+			}
 		}catch(Exception e) {
 			VRAgentController.LOGGER.error("Exception; reason was:", e);
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
